Move table to correct import path
