+++ vizro-core/src/vizro/tables/__init__.py
@@ -1,4 +1,4 @@
-from vizro.charts.tables.dash_table import dash_data_table
+from vizro.tables.dash_table import dash_data_table
 
 # Please keep alphabetically ordered
 __all__ = ["dash_data_table"]
