+++ vizro-core/src/vizro/tables/dash_table.py
